backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for dream team project"
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "azure-identity>=1.19.0",
    "python-dotenv>=1.0.1",
    "autogen-agentchat==0.4.1",
    "autogen-ext[azure]==0.4.1",
    "autogen-ext[docker]==0.4.1",
    "autogen-core==0.4.1",
    "streamlit==1.41.1",
    "promptflow-tracing==1.17.1",
    "azure-search-documents==11.6.0b4",
    "markitdown==0.0.1a3",
    "aiofiles==24.1.0",
    "aiohttp==3.11.11",
    "playwright>=1.49.1",
    "fastapi",
    "azure-storage-blob",
    "uvicorn",
    "fastapi-cors",
    "python-multipart",
    "tiktoken",
    "jinja2",
    "azure-cosmos"

]
